Query-replace: Replace `tags-query-replace' obsolete use, fix warning
Starting from Emacs 27.1, calling `tags-query-replace' with the file list
argument has been obsoleted. A significant part of the multi-search/replace
functionality is now provided by 'fileloop.el' (and `tags-query-replace' works
on top of it). It seems it is the most appropriate to replace the old call to
`tags-query-replace'. Change in behavior in particular allows to simplify the
code a lot (no `condition-case' to catch a normal end of search is anymore
needed).

A warning specified that the 'message' parameter was not used, and that was
what prompted its rename into '_message'. But this was a mistake: The warning
was effectively describing the actual situation. The value of '_message' was
indeed not passed to `sly-eval' at all (in fact, the whole enclosing `mapcar'
form was not executed) because of a misquote. This is now fixed.