Remove coupling between sly-asdf proper and flymake's support (#35)
Reason for doing this: Be able to ship just the part that depends on sly
proper, and nothing else (outside of Emacs' core packages), i.e., not on
Flymake nor popup.

To achieve this, I removed the `require' on sly-asdf-flymake in sly-asdf, as
well as calls to the former's functions (moved to the latter). Some functions
and variables from sly-asdf-flymake had to be moved into sly-asdf.

From users' point of view, the only thing that changes is that they need to
explicitly `require' sly-asdf-flymake, in addition to setting
`sly-asdf-enable-experimental-syntax-checking' to t (this variable could now be
removed).

I left the dependency on popup in 'sly-asdf.el'. Maybe the Flymake part can be
put in a separate package. Flymake should probably be added as an explicit
dependency of sly-asdf or of the new package.

Co-authored-by: Olivier Certner <olce.emacs@certner.fr>